refactor of kepler propagator
