+++ examples/run_full_iod_parallel.rs
@@ -61,6 +61,8 @@ fn outfit_error_label(err: &OutfitError) -> &'static str {
         OutfitError::DifferentialCorrectionFailed(_) => "DiffCorFailed",
         OutfitError::EphemerisBodyNotSupported(_) => "EphemerisBodyNotSupported",
         OutfitError::NBodyPropagationFailed(_) => "NBodyPropagationFailed",
+        OutfitError::DegenerateState(_) => "DegenerateState",
+        OutfitError::KeplerConvergence => "KeplerConvergence",
     }
 }
