+++ src/outfit_errors.rs
@@ -284,6 +284,12 @@ pub enum OutfitError {
 
     #[error("N-body propagation failed: {0}")]
     NBodyPropagationFailed(String),
+
+    #[error("{0}")]
+    DegenerateState(String),
+
+    #[error("universal Kepler solver did not converge")]
+    KeplerConvergence,
 }
 
 impl From<&ObsDatasetError> for OutfitError {
